Exploding Bullets
Bullets can now make explosions, but those explosions are currently only
at position 0,0,0. Still need to work on that.

+++ BugLogic/Damage.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Damage : MonoBehaviour {
 
-	public int damage;
+	public float damage;
 
 	// Use this for initialization
 	void Start () {
